reorganize classes and add new icon
