+++ src/main/java/net/alephsmp/aleph_carpet/mixins/endMainIslandStructureGen/FightManagerMixin.java
