+++ src/main/java/net/alephsmp/aleph_carpet/mixins/noCooldown/EndGatewayBlockEntityMixin.java
@@ -1,4 +1,4 @@
-package net.alephsmp.aleph_carpet.mixins;
+package net.alephsmp.aleph_carpet.mixins.noCooldown;
 
 import net.alephsmp.aleph_carpet.AlephSimpleSettings;
 import net.minecraft.block.entity.EndGatewayBlockEntity;
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
 @Mixin(EndGatewayBlockEntity.class)
-public abstract class EndGatewayBlockEntity_noCooldownMixin extends EndPortalBlockEntity {
+public abstract class EndGatewayBlockEntityMixin extends EndPortalBlockEntity {
     @Redirect(method="method_30276", at=@At(value="INVOKE", target="Lnet/minecraft/entity/Entity;hasNetherPortalCooldown()Z"))
     private static boolean returnFalseIfNoCooldown(Entity entity) {
         if (!AlephSimpleSettings.endGatewayCooldown) {
