+++ src/main/java/net/alephsmp/aleph_carpet/mixins/noCooldown/EndGatewayMixin.java
@@ -1,4 +1,4 @@
-package net.alephsmp.aleph_carpet.mixins;
+package net.alephsmp.aleph_carpet.mixins.noCooldown;
 
 import net.alephsmp.aleph_carpet.AlephSimpleSettings;
 import net.minecraft.block.entity.EndGatewayBlockEntity;
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(EndGatewayBlockEntity.class)
-public abstract class EndGateway_noCooldownMixin extends EndPortalBlockEntity {
+public abstract class EndGateway extends EndPortalBlockEntity {
     @Inject(method="needsCooldownBeforeTeleporting", at=@At("RETURN"), cancellable = true)
     private void suppressCooldown(CallbackInfoReturnable<Boolean> cir) {
         if (!AlephSimpleSettings.endGatewayCooldown) {
